Implemented spatial observer profiles and data providers
