+++ Assets/MixedRealityToolkit-SDK/Profiles/DefaultMixedRealitySpatialAwarenessProfile.asset
@@ -11,33 +11,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 07ce77c4cbff53f458a4a58f1407173a, type: 3}
   m_Name: DefaultMixedRealitySpatialAwarenessProfile
   m_EditorClassIdentifier: 
-  isCustomProfile: 0
-  spatialObserverDataProviders: {fileID: 11400000, guid: 4fd85ea3a7088a14d8375064a04233a7,
-    type: 2}
-  startupBehavior: 0
-  observationExtents: {x: 3, y: 3, z: 3}
-  isStationaryObserver: 0
-  updateInterval: 3.5
-  useMeshSystem: 1
-  meshPhysicsLayer: 31
-  meshLevelOfDetail: 0
-  meshTrianglesPerCubicMeter: 0
-  meshRecalculateNormals: 1
-  meshDisplayOption: 1
-  meshVisibleMaterial: {fileID: 2100000, guid: 47c3d3b0d8143e3489351498fceed55d, type: 2}
-  meshOcclusionMaterial: {fileID: 2100000, guid: a809c7ea182b699409443d46d5b42435,
-    type: 2}
-  useSurfaceFindingSystem: 0
-  surfaceFindingPhysicsLayer: 30
-  surfaceFindingMinimumArea: 0.025
-  displayFloorSurfaces: 0
-  floorSurfaceMaterial: {fileID: 2100000, guid: 47f3c5e1cb6142ba9697cd4c86d74321,
-    type: 2}
-  displayCeilingSurfaces: 0
-  ceilingSurfaceMaterial: {fileID: 2100000, guid: d5334c45caee46be937b095a1e977dc6,
-    type: 2}
-  displayWallSurfaces: 0
-  wallSurfaceMaterial: {fileID: 2100000, guid: 3c55769e893c4f4c8c51b7fa69bee2b9, type: 2}
-  displayPlatformSurfaces: 0
-  platformSurfaceMaterial: {fileID: 2100000, guid: cb74390ea57642f8b5ca0aa8f5fb38c1,
-    type: 2}
+  isCustomProfile: 1
+  registeredSpatialObserverDataProviders:
+  - spatialObserverType:
+      reference: Microsoft.MixedReality.Toolkit.Core.DataProviders.SpatialObservers.WindowsMixedReality.WindowsMixedRealitySpatialMeshObserver,
+        Microsoft.MixedReality.Toolkit.Core.Devices.SpatialObservers.WindowsMixedReality
+    spatialObserverName: Windows Mixed Reality Spatial Mesh Observer
+    priority: 5
+    runtimePlatform: 8
+    profile: {fileID: 11400000, guid: 25a015577dafdb245a9428581598f4d6, type: 2}
