--- Assets/MixedRealityToolkit-SDK/Profiles/DefaultMixedRealitySpatialObserverDataProvidersProfile.asset
@@ -1,22 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c4c54f1a09f94fd49ef7c8f0df3be5cb, type: 3}
-  m_Name: DefaultMixedRealitySpatialObserverDataProvidersProfile
-  m_EditorClassIdentifier: 
-  isCustomProfile: 0
-  registeredSpatialObserverDataProviders:
-  - spatialObserverType:
-      reference: Microsoft.MixedReality.Toolkit.Core.DataProviders.SpatialObservers.WindowsMixedReality.WindowsMixedRealitySpatialMeshObserverDataProvider,
-        Microsoft.MixedReality.Toolkit.Core.Devices.SpatialObservers.WindowsMixedReality
-    spatialObserverName: Windows Mixed Reality Spatial Observer Data Provider
-    priority: 5
-    runtimePlatform: 8
-    profile: {fileID: 11400000, guid: 47999fb2978b3554f95ea8ad2a51c0e2, type: 2}
