+++ Assets/MixedRealityToolkit-SDK/Profiles/SpatialObservers/WindowsMixedRealitySpatialMeshObserverProfile.asset
@@ -1,0 +1,26 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 90a4bdbfdea331745b57570c107eeaae, type: 3}
+  m_Name: WindowsMixedRealitySpatialMeshObserverProfile
+  m_EditorClassIdentifier: 
+  isCustomProfile: 1
+  startupBehavior: 0
+  observationExtents: {x: 3, y: 3, z: 3}
+  isStationaryObserver: 1
+  updateInterval: 3.5
+  physicsLayer: 31
+  meshLevelOfDetail: 0
+  meshTrianglesPerCubicMeter: 0
+  meshRecalculateNormals: 1
+  meshDisplayOption: 1
+  meshVisibleMaterial: {fileID: 2100000, guid: 47c3d3b0d8143e3489351498fceed55d, type: 2}
+  meshOcclusionMaterial: {fileID: 2100000, guid: a809c7ea182b699409443d46d5b42435,
+    type: 2}
