+++ Assets/MixedRealityToolkit-SDK/Profiles/SpatialObservers/WindowsMixedRealitySpatialMeshObserverProfile.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47999fb2978b3554f95ea8ad2a51c0e2
+guid: 25a015577dafdb245a9428581598f4d6
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
