+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/BaseMixedRealitySpatialMeshObserver.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: b4a405d6f4254e88b93f1d06d8314a43
+guid: 4764a686c33b4525a41f7521c592f9d4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {fileID: 2800000, guid: 3d652c19bde24a29871839929991b0be, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
