+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/BaseMixedRealitySpatialObserverDataProvider.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 6905e29a6bb04f3fb14e0eb8a689db7c
+guid: 331340f6dbc56884a81b44ed00b66dea
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
+  icon: {fileID: 2800000, guid: 3d652c19bde24a29871839929991b0be, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
