--- Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/CustomMixedRealitySpatialObserverProfile.cs
@@ -1,14 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-
-namespace Microsoft.MixedReality.Toolkit.Core.DataProviders.SpatialObservers
-{
-    // No Scriptable Object Menu constructor attributes here, as this class is meant to be inherited.
-
-    /// <summary>
-    /// Custom Mixed Reality Observer Profile class to inherit from to enable custom profile support.
-    /// </summary>
-    public abstract class CustomMixedRealitySpatialObserverProfile : BaseMixedRealitySpatialObserverProfile
-    {
-    }
-}
