+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/Profiles.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 4fd85ea3a7088a14d8375064a04233a7
-NativeFormatImporter:
+guid: d6d936a4b3f951c41b4b58c3e382c808
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
