+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/Profiles/BaseMixedRealitySpatialMeshObserverProfile.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: d601127b147144c4b4c8b6a3fe251f1f
+guid: 183894c01fba425094d7970230f31660
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
