+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/Profiles/BaseMixedRealitySpatialObserverProfile.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: fbddf422de884f8fab59357c4e8fa9b0
+guid: b4a405d6f4254e88b93f1d06d8314a43
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
