+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/Profiles/BaseMixedRealitySurfaceObserverProfile.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 331340f6dbc56884a81b44ed00b66dea
+guid: c387bdb733d34c17b55c8c45cb19580b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
