+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/WindowsMixedReality/Profiles/Inspectors/WindowsMixedRealitySpatialMeshObserverProfileInspector.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Core.Inspectors.Profiles;
+using Microsoft.MixedReality.Toolkit.Core.Inspectors.Profiles.SpatialAwareness;
 using UnityEditor;
 
 namespace Microsoft.MixedReality.Toolkit.Core.DataProviders.SpatialObservers.WindowsMixedReality.Profiles.Inspectors
 {
     [CustomEditor(typeof(WindowsMixedRealitySpatialMeshObserverProfile))]
-    public class WindowsMixedRealitySpatialMeshObserverProfileInspector : BaseMixedRealityInspector { }
+    public class WindowsMixedRealitySpatialMeshObserverProfileInspector : BaseMixedRealitySpatialMeshObserverProfileInspector { }
 }
