+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/WindowsMixedReality/Profiles/WindowsMixedRealitySpatialMeshObserverProfile.cs
@@ -1,13 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Core.Definitions;
 using UnityEngine;
 
 namespace Microsoft.MixedReality.Toolkit.Core.DataProviders.SpatialObservers.WindowsMixedReality.Profiles
 {
     [CreateAssetMenu(menuName = "Mixed Reality Toolkit/Spatial Observers/Windows Mixed Reality Spatial Mesh Data Provider Profile", fileName = "WindowsMixedRealitySpatialMeshObserverProfile")]
-    public class WindowsMixedRealitySpatialMeshObserverProfile : BaseMixedRealitySpatialObserverProfile
-    {
-    }
+    public class WindowsMixedRealitySpatialMeshObserverProfile : BaseMixedRealitySpatialMeshObserverProfile { }
 }
