+++ Assets/MixedRealityToolkit/_Core/DataProviders/SpatialObservers/WindowsMixedReality/WindowsMixedRealitySpatialMeshObserver.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 695870a3b55a4b5ca6885049aea1921f
+guid: 3f270166d6a6c1a4099fe4e6163d4713
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
+  icon: {fileID: 2800000, guid: 3d652c19bde24a29871839929991b0be, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
