--- Assets/MixedRealityToolkit/_Core/Definitions/Devices/SpatialObserverType.cs
@@ -1,15 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-
-namespace Microsoft.MixedReality.Toolkit.Core.Definitions.Devices
-{
-    /// <summary>
-    /// The types of spatial observers supported by the Mixed Reality Toolkit.
-    /// </summary>
-    public enum SpatialObserverType
-    {
-        None = 0,
-        Mesh,
-        Surface
-    }
-}
