+++ Assets/MixedRealityToolkit/_Core/EventDatum/SpatialAwareness/MixedRealitySpatialAwarenessEventData.cs
@@ -33,7 +33,7 @@ namespace Microsoft.MixedReality.Toolkit.Core.EventDatum.SpatialAwarenessSystem
         /// <param name="spatialAwarenessObserver"></param>
         /// <param name="id"></param>
         /// <param name="spatialObject"></param>
-        public void Initialize(IMixedRealitySpatialAwarenessObserver spatialAwarenessObserver, int id, T spatialObject)
+        public void Initialize(IMixedRealitySpatialObserverDataProvider spatialAwarenessObserver, int id, T spatialObject)
         {
             BaseInitialize(spatialAwarenessObserver);
             Id = id;
