+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialObservers.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: f6501baa30c0b704595cec21159c23a5
-NativeFormatImporter:
+guid: 9d3f310959bc2f546a8a642a01a8d183
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
