+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialObservers/BaseMixedRealitySpatialMeshObserverProfileInspector.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 71f25d32570fe144fb2cd41b61c98393
+guid: 048516155ded431a8f84c66b8a259223
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
