+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialObservers/BaseMixedRealitySpatialObserverProfileInspector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 695870a3b55a4b5ca6885049aea1921f
+guid: 4b7374a6b79242c7998ead9733df9dc7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
