+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialObservers/BaseMixedRealitySurfaceObserverProfileInspector.cs
@@ -1,0 +1,55 @@
+﻿using UnityEditor;
+using UnityEngine;
+
+namespace Microsoft.MixedReality.Toolkit.Core.Inspectors.Profiles.SpatialAwareness
+{
+    public abstract class BaseMixedRealitySurfaceObserverProfileInspector : BaseMixedRealitySpatialObserverProfileInspector
+    {
+        private SerializedProperty surfaceFindingMinimumArea;
+        private SerializedProperty displayFloorSurfaces;
+        private SerializedProperty floorSurfaceMaterial;
+        private SerializedProperty displayCeilingSurfaces;
+        private SerializedProperty ceilingSurfaceMaterial;
+        private SerializedProperty displayWallSurfaces;
+        private SerializedProperty wallSurfaceMaterial;
+        private SerializedProperty displayPlatformSurfaces;
+        private SerializedProperty platformSurfaceMaterial;
+
+        private readonly GUIContent ceilingMaterialContent = new GUIContent("Ceiling Material");
+        private readonly GUIContent floorMaterialContent = new GUIContent("Floor Material");
+        private readonly GUIContent platformMaterialContent = new GUIContent("Platform Material");
+        private readonly GUIContent wallMaterialContent = new GUIContent("Wall Material");
+        private readonly GUIContent minimumAreaContent = new GUIContent("Minimum Area");
+
+        /// <inheritdoc />
+        protected override void OnEnable()
+        {
+            base.OnEnable();
+
+            surfaceFindingMinimumArea = serializedObject.FindProperty("surfaceFindingMinimumArea");
+            displayFloorSurfaces = serializedObject.FindProperty("displayFloorSurfaces");
+            floorSurfaceMaterial = serializedObject.FindProperty("floorSurfaceMaterial");
+            displayCeilingSurfaces = serializedObject.FindProperty("displayCeilingSurfaces");
+            ceilingSurfaceMaterial = serializedObject.FindProperty("ceilingSurfaceMaterial");
+            displayWallSurfaces = serializedObject.FindProperty("displayWallSurfaces");
+            wallSurfaceMaterial = serializedObject.FindProperty("wallSurfaceMaterial");
+            displayPlatformSurfaces = serializedObject.FindProperty("displayPlatformSurfaces");
+            platformSurfaceMaterial = serializedObject.FindProperty("platformSurfaceMaterial");
+        }
+
+        /// <inheritdoc />
+        public override void OnInspectorGUI()
+        {
+            EditorGUILayout.LabelField("Surface Finding Settings:", EditorStyles.boldLabel);
+            EditorGUILayout.PropertyField(surfaceFindingMinimumArea, minimumAreaContent);
+            EditorGUILayout.PropertyField(displayFloorSurfaces);
+            EditorGUILayout.PropertyField(floorSurfaceMaterial, floorMaterialContent);
+            EditorGUILayout.PropertyField(displayCeilingSurfaces);
+            EditorGUILayout.PropertyField(ceilingSurfaceMaterial, ceilingMaterialContent);
+            EditorGUILayout.PropertyField(displayWallSurfaces);
+            EditorGUILayout.PropertyField(wallSurfaceMaterial, wallMaterialContent);
+            EditorGUILayout.PropertyField(displayPlatformSurfaces);
+            EditorGUILayout.PropertyField(platformSurfaceMaterial, platformMaterialContent);
+        }
+    }
+}
