+++ Assets/MixedRealityToolkit/_Core/Inspectors/Profiles/SpatialObservers/BaseMixedRealitySurfaceObserverProfileInspector.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 3f270166d6a6c1a4099fe4e6163d4713
+guid: 46f9add87be84f84a60f1020949bf159
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
