+++ Assets/MixedRealityToolkit/_Core/Interfaces/DataProviders/SpatialObservers/IMixedRealitySpatialMeshObserver.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 28f2096fc29a45269d694224abb49b22
+timeCreated: 1544047414
