+++ Assets/MixedRealityToolkit/_Core/Interfaces/DataProviders/SpatialObservers/IMixedRealitySpatialObserverDataProvider.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: c4c54f1a09f94fd49ef7c8f0df3be5cb
+guid: a843c5157f6af4846bc7888caf4c9c3f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 4f9f54f9478441228dea18a2c828cfc6, type: 3}
+  icon: {instanceID: 0}
   userData: 
   assetBundleName: 
   assetBundleVariant:
