+++ Assets/MixedRealityToolkit/_Core/Interfaces/DataProviders/SpatialObservers/IMixedRealitySurfaceObserver.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7fb96030d5d846639cc0d2d6994eda4e
+timeCreated: 1544047418
