+++ Assets/MixedRealityToolkit/_Core/Services/MixedRealityToolkit.cs
@@ -208,13 +208,19 @@ namespace Microsoft.MixedReality.Toolkit.Core.Services
 #endif
                 if (RegisterService<IMixedRealitySpatialAwarenessSystem>(ActiveProfile.SpatialAwarenessSystemSystemType) && SpatialAwarenessSystem != null)
                 {
-                    if (ActiveProfile.SpatialAwarenessProfile.SpatialObserverDataProviders != null &&
-                        ActiveProfile.SpatialAwarenessProfile.SpatialObserverDataProviders.RegisteredSpatialObserverDataProviders != null)
+                    if (ActiveProfile.SpatialAwarenessProfile.RegisteredSpatialObserverDataProviders != null)
                     {
-                        for (int i = 0; i < ActiveProfile.SpatialAwarenessProfile.SpatialObserverDataProviders.RegisteredSpatialObserverDataProviders.Length; i++)
+                        for (int i = 0; i < ActiveProfile.SpatialAwarenessProfile.RegisteredSpatialObserverDataProviders.Length; i++)
                         {
-                            var spatialObserver = ActiveProfile.SpatialAwarenessProfile.SpatialObserverDataProviders.RegisteredSpatialObserverDataProviders[i];
-                            RegisterService<IMixedRealitySpatialAwarenessObserver>(spatialObserver.SpatialObserverType, spatialObserver.RuntimePlatform, spatialObserver.SpatialObserverName, spatialObserver.Priority);
+                            var spatialObserver = ActiveProfile.SpatialAwarenessProfile.RegisteredSpatialObserverDataProviders[i];
+
+                            if (spatialObserver.Profile == null)
+                            {
+                                Debug.LogError($"Missing profile for {spatialObserver.SpatialObserverName}");
+                                continue;
+                            }
+
+                            RegisterService<IMixedRealitySpatialObserverDataProvider>(spatialObserver.SpatialObserverType, spatialObserver.RuntimePlatform, spatialObserver.SpatialObserverName, spatialObserver.Priority, spatialObserver.Profile);
                         }
                     }
                 }
